==> scr/00_libraries.R <==
renv::install("tidyverse")
renv::install("bioc::DESeq2")
renv::install("bioc::pasilla")
renv::install("bioc::limma")
renv::install("bioc::vsn")
renv::install("hexbin")
renv::install("viridis")
renv::install("pheatmap",rebuild = T)
renv::install("PoiClaClu")
renv::install("ashr")
renv::install("contrast")
renv::install("bioc::AnnotationDbi") 
renv::install("bioc::AnnotationHub")
renv::install("bioc::ensembldb")
renv::install("ggrepel")
renv::install("bioc::IRanges")
renv::install("jokergoo/ComplexHeatmap")
renv::install("bioc::fgsea")
renv::install("bioc::msigdbr")
renv::install("bioc::GSEABase")
renv::install("bioc::patchwork")
renv::install("UpSetR")
renv::install("gplots")
renv::install("DT")
renv::install("GGally")
renv::install("corrplot")
renv::install("RNAseqQC")
renv::install("enrichR")

renv::install("matrixStats@1.1.0")
renv::install("bioc::org.Hs.eg.db")
renv::install("bioc::GSVA")
